Stub for tbtc-dapp deployment
A few assumptions here:  That we don't need an InitContainer, the config
is sparse and there are no contract address configs.  We can hardcode
the config for now.  We're not exposing the service to the internet so
it's just a ClusterIP.  Replica count is 1 for now, don't think we need
an HA deployment for a demo.  Port is not set, not sure if it's just
standard http or if there's a defined port.  Will check in a bit.