Function to broadcast a transaction via electrum
Adds a broadcastTransaction function to Electrum Client. The function
takes raw transaction provided as a hexadecimal string and forwards it
to electrum network in blockchain.transaction.broadcast method.
As a result it returns hash of the submitted transaction or throws an
error in case of failure.

See: https://electrumx.readthedocs.io/en/latest/protocol-methods.html#blockchain-transaction-broadcast