Update MarkeProof verification
We can't validate Merkle with Root and hash extracted from proof
since they are no longer provided in the proof.
We pass the txHash directly, convert it into a buffer and reverse endianes.
We then use provided Block height to get the expected Root from client.
If the derived root equals the root returned by client, we return true