Demo script for redemption transaction
The scripts executes steps of creating and submitting a redemption
transaction. It obtains required data from events logged on chain and
tries to recreate a transaction the same way as bitcoin-spv does, so the
digest used for signing by tBTC contract matches the one we create it in
this script.