Some work on regtest support
Currently there's a local ElectrumX server, connected to a regtest testnet. For some reason, bitcoin-cli doesn't support bech32 on regtest (complaining it's an invalid from). Transacting works with a regular base58 address.

The current issue is ElectrumX is not picking up the funding transaction to this base58 address. It might be due to the script hash, but it could also be a different network version, I'm not sure.