Client scripts for redemption
Implemented functions to redeem a deposit. First it calls the deposit
to request a redemption. Later finds events logged on chain to obtain
data required to create a redemption transaction and a signature
calculated by a keep. Finally it puts the found data together to form
a bitcoin transaction and broadcasts the transaction to the bitcoin chain.