Rewrite getDepositBtcAddress to try existing logs
getDepositBtcAddress was set up to always submit a retrieveSignerPubkey
transaction to the chain in order to trigger a BTC address on-chain log;
only after submitting this transaction did the wait for a log begin.

Since this function will be used for restoring deposit state as well,
not just for linearly progressing, this commit rearranges the logic to
start with an attempt to read an existing log, and only submit the
transaction if the log does not already exist.