Introduce deposit.stateRestored in redux state
This state is updated when the RESTORE_DEPOSIT_STATE saga is finished
doing whatever it needs to do to get the deposit state booted back up
when the page loads from scratch in mid-flow. Currently it is
immediately set when a deposit awaiting funding proof loads up, though
some state restoration code needs to be inserted.

The state is used to clear the loading component that is presented on
mid-flow states.