Rework scripts/fetch-contracts.sh to run from root
It was trying to guess at where the script itself was, then trying to
adjust the directory, but this was done in a non-portable way. The
script is now explicit about expecting to be run from repo root, and is
also explicit about expecting the right directory structure to exist
before it pulls down contracts.