Add hacky Web3Loaded promise for web3 wrapper
This promise is controlled by an unexported-but-global variable that
gets updated by the Web3Wrapper, and that is polled every 100ms for
loading to have completed. This is, in fact, terrible, and the entire
mechanism needs replacement with a saga that handles web3 connections
and updates the redux state with the resulting information, but for now
it's sufficient to build sagas that depends on the web3 loading state.