Give redemption proper URLs and Loadable support
Loadable is now parametrized via a "restorer" parameter, which only
accepts values "redemption" and "deposit" for now. This dispatches to
the appropriate restore function, re-hydrating the state in the
appropriate state subset. Existing Loadable wrappers are
Loadable restorer="deposit", while the redemption flows are now
Loadable restorer="redemption".

Redemption addresses are in the form:

    /deposit/<address>/redemption[/<step>]

In this sense, "redemption" is a singleton subresource of the deposit.