Add basic address validation in redemption flow
Right now this is fairly basic, and the tBTC helpers don't expose a
helper function for validation. This should be pulled into the tBTC
helper library so that the deposit address is not only checked for
being a valid Ethereum address, but also a valid deposit. Similarly, the
Bitcoin address can be checked for any new address types.

No errors are currently displayed to the user beyond showing that the
address as inputted is invalid.