FIXME a niggle that we'll have to deal with
In cases where someone is loading a page that is on the wrong page state
and that page state generates an Ethereum transaction, even though the
final page and URL are correct, the original page's component's
componentDidMount still runs, so it can still fire a now-stale
transaction. That needs to be attended to, but is a corner case for now.