Support arbitrary redemption script addresses
The filtering here is fairly simple, using bcoin.Script.fromAddress and
expecting that (1) an exception will be thrown for invalid addresses and
(2) getAddress will return null if any non-exception worthy results come
out of Script.fromAddress.

The chain requires a length prefix for the resulting script, so the raw
script has its own length prepended before it is shipped to the deposit
contract.