Grossly adjust web3 wrapper to also load TBTC
While there are far cleaner ways to do this, the dirtiness is already
done in the web3 wrapper. A better home will be needed for TBTC once
web3 loading is fixed, but in the meantime the hack gets some hack
tacked atop it. The exported TBTCLoaded property provides a promise to a
fully-configured TBTC object.

In the process, contract initialization falls out, as the TBTC
initialization takes care of contract initialization as well when using
tbtc.js.