Move deposit flow to tbtc.js
This largely involves deleting code. The deposit saga now consists of
two sagas, one for requesting a deposit and one for auto-submitting
deposit proof. Waiting for a confirmation is handled internally in the
auto-submitting saga.

Additionally, handling of BTC amounts is in the view, while sagas and
the redux store deal entirely in satoshis.

Last but not least, the DEPOSIT_RESOLVED action and reducer are used to
signal the readiness of a tbtc.js Deposit object, and to store it in the
redux store.