Migrate to default namespace, use proper cert
There's a couple things going on here.  First, we're shipping the
tbtc-dapp and it's services to the default namespace.  Frankly I was
trying to be cute when initially shipping tBTC to our internal networks
and it's just extra overhead that's not needed.  We're a small ship.

Second, we're fronting our dapp deployment with a proper cert from
Cloudflare.  I should have done this ages ago.