Resume at /confirming for AWAITING_WITHDRAWAL_PROOF
This state was pointing to /redemption/prove, but in fact just because
the deposit is waiting for proof, doesn't mean it's already had its
proper number of confirmations. The deposit restorer now instead
redirects to /confirming, which will transition to /prove when the right
number of confirmations have been received---immediately if they already
exist.