Update Confirming redemption view to better reflect status
The old logic expected that we would know the current number of confirmations,
but the tbtc.js currently only exposes whether confirmations are still pending or
completed. Also, by this step we already finished broadcasting the signed
transaction so this removes that copy.

This also removes unused actions and cases in the redemption sagas and
reducer.