Adjust app max-width to be wider
This allows the lot size selector to appear on one row for larger screens.
Previously the max-width for the whole app was 768px. Makes exceptions
for the home page, since the home page looks weird if you try to
extend beyond the original 768px.

Also adds some mixins and new breakpoint variables and moves the layout
logic from app.scss to structure.scss