When handling errors, deal with the possible reason property
Web3 errors tend to encode a contract's revert reason as a reason
property, independent of the message property (which doesn't mention the
revert reason at all at the moment). logError now includes that reason
in the console.log call, and uses it as the payload for the error action
if it exists.