Move publishing of Go client to GH Actions
As described in RFC-18, there is a need for a refactorization of Keep
and tBTC release processes in order to reduce human involvement and to
make the process less error prone. A part of this task is migration
from CircleCI to GitHub Actions. This commit creates a GitHub Action
workflow for linting, building and publishing of tBTC Dapp.

The `build-and-publish` job is configured to execute only if
workflow is started by a push or workflow_dispatch event on master or
`rfc-18/*` branch. The `rfc-18/*` condition will need to be removed
once whole `build-test-migrate-publish-keep-test` CircleCI process gets
migrated to GH Actions.