Cleanup after moving to GH Actions
As described in RFC-18, there is a need for a refactorization of Keep
and tBTC release processes in order to reduce human involvement and
make the processes faster and less error prone. A part of this task is
migration from CircleCI jobs to GitHub Actions.

Now that we have all CircleCI jobs from `tbtc-dapp` moved to GH Actions
(and verified), we can implement some final changes removing obsolete
CircleCI configs and cleaning up pieces of code used for workflows
testing. Here is a list of changes applied:
- removed unnecessary comments
- deleted `./circleci/config.yml` (CircleCI jobs are no longer needed)
- removed memory monitoring step in client workflows
- removed references to `rfc-18...` branches, which were used for
  workflow testing
- removed `-wip` suffix from name of the published image