Fix routing
For some reason, `Route` components are not wrapped with the` Switch`
component. In that case not match route:
```
<Rout path="*">
 <Redirect to="/" />
</Route>
```
doesn't work as expected and you can't go to a given page direectly eg.
`redeem`. The dapp always redirects to home page(`/`).

Because I don't know the reason why we don't use `Switch` component and
to avoid crash, this commit redirects from `/deposit` and `deposit/new`
to a home page(`/`) as a workaround.